Add end-to-end diff test
Creates a test which performs a full xml diff against the input file and the saved copy of the input file.  Eventually, there will be no differences.  When running the test, you can open up the diff html to visually see what is missing or changed from the source file.
Added XmlDiffPatch as nuget dependency.  This library is used to compare the xml files and generate the visual diff in html.